File rearrangement: 1) Rename screen into ia. 2) Move UI into framework
