+++ src/main/java/abr/user_avatar_image_management_abr/UserAvatarMngUseCase.java
@@ -1,6 +1,6 @@
 package abr.user_avatar_image_management_abr;
 
-import interface_adaptors.user_avatar_image_management_screen.UserAvatarMngViewModel;
+import interface_adaptors.user_avatar_image_management_ia.UserAvatarMngViewModel;
 import ds.user_avatar_image_management_ds.UserAvatarDatabaseGateway;
 import entities.user_entities.User;
 import entities.user_entities.UserAvatar;
