+++ src/main/java/abr/user_login_abr/UserLogOutputBoundary.java
@@ -1,6 +1,6 @@
 package abr.user_login_abr;
 
-import interface_adaptors.user_login_screen.UserLogViewModel;
+import interface_adaptors.user_login_ia.UserLogViewModel;
 
 public interface UserLogOutputBoundary {
     UserLogViewModel packageAndPresent(UserLogResponseModel responseModel);
