+++ src/main/java/abr/user_reg_abr/UserRegInputBoundary.java
@@ -1,6 +1,6 @@
 package abr.user_reg_abr;
 
-import interface_adaptors.user_reg_screen.UserRegViewModel;
+import interface_adaptors.user_reg_ia.UserRegViewModel;
 
 public interface UserRegInputBoundary {
     UserRegViewModel register(UserRegRequestModel requestModel);
