+++ src/main/java/abr/user_reg_abr/UserRegOutputBoundary.java
@@ -1,6 +1,6 @@
 package abr.user_reg_abr;
 
-import interface_adaptors.user_reg_screen.UserRegViewModel;
+import interface_adaptors.user_reg_ia.UserRegViewModel;
 
 public interface UserRegOutputBoundary {
     UserRegViewModel packageAndPresent(UserRegResponseModel responseModel);
