+++ src/main/java/interface_adaptors/user_avatar_image_management_ia/UserAvatarMngController.java
@@ -1,4 +1,4 @@
-package interface_adaptors.user_avatar_image_management_screen;
+package interface_adaptors.user_avatar_image_management_ia;
 
 import abr.user_avatar_image_management_abr.UserAvatarMngInputBoundary;
 import abr.user_avatar_image_management_abr.UserAvatarMngRequestModel;
