+++ src/main/java/interface_adaptors/user_login_ia/UserLogController.java
@@ -1,4 +1,4 @@
-package interface_adaptors.user_login_screen;
+package interface_adaptors.user_login_ia;
 
 import abr.user_login_abr.UserLogInputBoundary;
 import abr.user_login_abr.UserLogRequestModel;
