+++ src/main/java/interface_adaptors/user_login_ia/UserLogViewModel.java
@@ -1,4 +1,4 @@
-package interface_adaptors.user_login_screen;
+package interface_adaptors.user_login_ia;
 
 public class UserLogViewModel {
     boolean userNameValid;
