+++ src/test/java/user_unit_test/login/UserLogUseCaseUnitTest.java
@@ -8,12 +8,12 @@ import ds.user_reg_ds.UserRegisterFileGateway;
 import abr.user_login_abr.UserLogOutputBoundary;
 import abr.user_login_abr.UserLogRequestModel;
 import abr.user_login_abr.UserLogUseCase;
-import interface_adaptors.user_login_screen.UserLogPresenter;
-import interface_adaptors.user_login_screen.UserLogViewModel;
+import interface_adaptors.user_login_ia.UserLogPresenter;
+import interface_adaptors.user_login_ia.UserLogViewModel;
 import abr.user_reg_abr.UserRegOutputBoundary;
 import abr.user_reg_abr.UserRegRequestModel;
 import abr.user_reg_abr.UserRegUseCase;
-import interface_adaptors.user_reg_screen.UserRegPresenter;
+import interface_adaptors.user_reg_ia.UserRegPresenter;
 
 import java.util.HashMap;
 import java.util.Map;
