+++ src/test/java/user_unit_test/reg/UserRegUseCaseUnitTest.java
@@ -6,8 +6,8 @@ import ds.user_reg_ds.UserRegisterFileGateway;
 import abr.user_reg_abr.UserRegOutputBoundary;
 import abr.user_reg_abr.UserRegRequestModel;
 import abr.user_reg_abr.UserRegUseCase;
-import interface_adaptors.user_reg_screen.UserRegPresenter;
-import interface_adaptors.user_reg_screen.UserRegViewModel;
+import interface_adaptors.user_reg_ia.UserRegPresenter;
+import interface_adaptors.user_reg_ia.UserRegViewModel;
 
 import java.util.HashMap;
 import java.util.Map;
